requirements.txt
absl-py==1.0.0
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1635819655903/work
argon2-cffi @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi_1636021455522/work
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
beautifulsoup4==4.10.0
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
cffi @ file:///opt/concourse/worker/volumes/live/976f8942-f51d-4f0e-7352-2a10f0820d0e/volume/cffi_1625814703974/work
charset-normalizer==2.0.6
cryptography==3.4.8
cycler==0.11.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.2
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1639976188178/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1635615662634/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1620913000384/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1638470588570/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1631590360471/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1637175406194/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1636510082894/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1640026479618/work
jupyter @ file:///Users/runner/miniforge3/conda-bld/jupyter_1637233342265/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1616560109969/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1636814410433/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1631590465624/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1635833678538/work
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1635844789148/work
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert_1631125953237/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1637122013008/work
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.3.5
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.3.2
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1635538335951/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.1
psycopg2==2.9.1
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
PyJWT==2.1.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1636757021002/work
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1636052713700/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python3-openid==3.2.0
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pyzmq @ file:///opt/concourse/worker/volumes/live/dbbc086d-2905-496b-54ee-08109e9ced8d/volume/pyzmq_1638435006316/work
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1639432716385/work
QtPy @ file:///home/conda/feedstock_root/build_artifacts/qtpy_1640200777318/work
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1636052339462/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1635819729986/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
widgetsnbextension @ file:///Users/runner/miniforge3/conda-bld/widgetsnbextension_1637174461614/work
wrapt==1.13.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
